refactor(symbolic): consolidate solver opts
