+++ crates/evm/symbolic/src/runtime/solver.rs
@@ -2,7 +2,7 @@ use super::*;
 
 mod hard_arith_fallback;
 mod monotonic_product;
-mod normalize;
+mod opt;
 
 use hard_arith_fallback::constraints_prefer_hard_arith_fallback_first;
 #[cfg(test)]
@@ -11,9 +11,13 @@ pub(crate) use hard_arith_fallback::hard_arith_fallback_model;
 #[cfg(test)]
 pub(crate) use monotonic_product::product_monotonic_unsat;
 use monotonic_product::product_monotonic_unsat_normalized;
-pub(crate) use normalize::normalize_constraints_for_solver;
+pub(crate) use opt::normalize_constraints_for_solver;
+use opt::{
+    constraint_cache_key, constraints_are_directly_unsat, sorted_bool_exprs_are_subset,
+    write_smt_assertions,
+};
 #[cfg(test)]
-pub(crate) use normalize::{normalize_bool_for_solver, normalize_expr_for_solver};
+pub(crate) use opt::{normalize_bool_for_solver, normalize_expr_for_solver};
 
 /// Errors that arise when parsing or constructing solver commands from configuration.
 #[derive(Debug, thiserror::Error)]
@@ -514,18 +518,18 @@ impl SmtLibSubprocessSolver {
 
     /// Caches a definitive normalized satisfiability result if the cache has room.
     fn cache_sat_result(&mut self, key: Vec<SymBoolExpr>, result: bool) {
-        if self.sat_cache.contains_key(&key)
-            || self.sat_cache.len() < SYMBOLIC_SOLVER_SAT_CACHE_MAX_ENTRIES
-        {
+        if let Some(cached) = self.sat_cache.get_mut(&key) {
+            *cached = result;
+        } else if self.sat_cache.len() < SYMBOLIC_SOLVER_SAT_CACHE_MAX_ENTRIES {
             self.sat_cache.insert(key, result);
         }
     }
 
     /// Caches a validated normalized model result if the cache has room.
     fn cache_model_result(&mut self, key: Vec<SymBoolExpr>, model: SymbolicModel) {
-        if self.model_cache.contains_key(&key)
-            || self.model_cache.len() < SYMBOLIC_SOLVER_MODEL_CACHE_MAX_ENTRIES
-        {
+        if let Some(cached) = self.model_cache.get_mut(&key) {
+            *cached = model;
+        } else if self.model_cache.len() < SYMBOLIC_SOLVER_MODEL_CACHE_MAX_ENTRIES {
             self.model_cache.insert(key, model);
         }
     }
@@ -590,148 +594,6 @@ impl SmtLibSubprocessSolver {
             }
         }
         result.output
-    }
-}
-
-/// Returns a structural key for normalized solver cache lookups.
-fn constraint_cache_key(constraints: &[SymBoolExpr]) -> Vec<SymBoolExpr> {
-    let mut key = Vec::with_capacity(constraints.len());
-    for constraint in constraints.iter().cloned() {
-        constraint.cache_key().push_cache_key_conjuncts(&mut key);
-    }
-    key.sort();
-    key.dedup();
-    key
-}
-
-/// Returns whether normalized conjunctive constraints contain a direct contradiction.
-fn constraints_are_directly_unsat(constraints: &[SymBoolExpr]) -> bool {
-    constraints.iter().any(|constraint| match constraint.kind() {
-        SymBoolExprKind::Const(false) => true,
-        SymBoolExprKind::Not(inner) => constraints.binary_search(inner).is_ok(),
-        _ => constraints.binary_search(&constraint.clone().not()).is_ok(),
-    })
-}
-
-/// Returns whether every expression in sorted `subset` appears in sorted `superset`.
-fn sorted_bool_exprs_are_subset(subset: &[SymBoolExpr], superset: &[SymBoolExpr]) -> bool {
-    if subset.len() > superset.len() {
-        return false;
-    }
-
-    let mut superset = superset.iter();
-    for expected in subset {
-        loop {
-            match superset.next() {
-                Some(candidate) if candidate < expected => {}
-                Some(candidate) if candidate == expected => break,
-                _ => return false,
-            }
-        }
-    }
-    true
-}
-
-impl SymBoolExpr {
-    fn cache_key(self) -> Self {
-        self.fold(&mut Self::cache_key_node)
-    }
-
-    fn cache_key_node(expr: Self) -> Self {
-        match expr.into_kind() {
-            SymBoolExprKind::Not(value) => value.not(),
-            SymBoolExprKind::And(values) => {
-                let mut conjuncts = Vec::new();
-                for value in values.iter().cloned() {
-                    value.push_cache_key_conjuncts(&mut conjuncts);
-                }
-                conjuncts.sort();
-                conjuncts.dedup();
-                Self::and(conjuncts)
-            }
-            SymBoolExprKind::Eq(left, right) => {
-                let left = left.cache_key();
-                let right = right.cache_key();
-                if left <= right { Self::eq(left, right) } else { Self::eq(right, left) }
-            }
-            SymBoolExprKind::Cmp(op, left, right) => {
-                Self::cache_key_cmp(op, left.cache_key(), right.cache_key())
-            }
-            SymBoolExprKind::Const(value) => Self::constant(value),
-        }
-    }
-
-    fn push_cache_key_conjuncts(self, out: &mut Vec<Self>) {
-        match self.kind() {
-            SymBoolExprKind::Const(true) => {}
-            SymBoolExprKind::And(values) => {
-                for value in values.iter().cloned() {
-                    value.push_cache_key_conjuncts(out);
-                }
-            }
-            _ => out.push(self),
-        }
-    }
-
-    fn cache_key_cmp(op: SymBoolExprOp, left: SymExpr, right: SymExpr) -> Self {
-        match op {
-            SymBoolExprOp::Ugt => Self::cmp(SymBoolExprOp::Ult, right, left),
-            SymBoolExprOp::Uge => Self::cmp(SymBoolExprOp::Ule, right, left),
-            SymBoolExprOp::Sgt => Self::cmp(SymBoolExprOp::Slt, right, left),
-            SymBoolExprOp::Ult | SymBoolExprOp::Ule | SymBoolExprOp::Slt => {
-                Self::cmp(op, left, right)
-            }
-        }
-    }
-}
-
-impl SymExpr {
-    fn cache_key(self) -> Self {
-        self.fold(&mut Self::cache_key_node)
-    }
-
-    fn cache_key_node(expr: Self) -> Self {
-        match expr.kind() {
-            SymExprKind::Op(op, left, right) => {
-                if op.is_commutative() && right < left {
-                    let SymExprKind::Op(op, left, right) = expr.into_kind() else { unreachable!() };
-                    Self::op(op, right, left)
-                } else {
-                    expr
-                }
-            }
-            SymExprKind::AddMod { left, right, .. } => {
-                if right < left {
-                    let SymExprKind::AddMod { left, right, modulus } = expr.into_kind() else {
-                        unreachable!()
-                    };
-                    Self::addmod(right, left, modulus)
-                } else {
-                    expr
-                }
-            }
-            SymExprKind::MulMod { left, right, .. } => {
-                if right < left {
-                    let SymExprKind::MulMod { left, right, modulus } = expr.into_kind() else {
-                        unreachable!()
-                    };
-                    Self::mulmod(right, left, modulus)
-                } else {
-                    expr
-                }
-            }
-            SymExprKind::Ite(_, _, _) => {
-                let SymExprKind::Ite(cond, left, right) = expr.into_kind() else { unreachable!() };
-                Self::ite(cond.cache_key(), left, right)
-            }
-            _ => expr,
-        }
-    }
-}
-
-impl SymExprOp {
-    const fn is_commutative(self) -> bool {
-        matches!(self, Self::Add | Self::Mul | Self::And | Self::Or | Self::Xor)
     }
 }
