+++ crates/evm/symbolic/src/runtime/solver/opt.rs
@@ -29,9 +29,54 @@ fn normalize_constraint_batch(
         }
         constraint.push_normalized_conjuncts(&mut normalized);
     }
-    normalized.sort();
-    normalized.dedup();
+    sort_dedup_bool_exprs(&mut normalized);
     normalized
+}
+
+fn sort_dedup_bool_exprs(exprs: &mut Vec<SymBoolExpr>) {
+    exprs.sort();
+    exprs.dedup();
+}
+
+/// Returns a structural key for normalized solver cache lookups.
+pub(super) fn constraint_cache_key(constraints: &[SymBoolExpr]) -> Vec<SymBoolExpr> {
+    let mut key = Vec::with_capacity(constraints.len());
+    for constraint in constraints.iter().cloned() {
+        constraint.cache_key().push_cache_key_conjuncts(&mut key);
+    }
+    sort_dedup_bool_exprs(&mut key);
+    key
+}
+
+/// Returns whether normalized conjunctive constraints contain a direct contradiction.
+pub(super) fn constraints_are_directly_unsat(constraints: &[SymBoolExpr]) -> bool {
+    constraints.iter().any(|constraint| match constraint.kind() {
+        SymBoolExprKind::Const(false) => true,
+        SymBoolExprKind::Not(inner) => constraints.binary_search(inner).is_ok(),
+        _ => constraints.binary_search(&constraint.clone().not()).is_ok(),
+    })
+}
+
+/// Returns whether every expression in sorted `subset` appears in sorted `superset`.
+pub(super) fn sorted_bool_exprs_are_subset(
+    subset: &[SymBoolExpr],
+    superset: &[SymBoolExpr],
+) -> bool {
+    if subset.len() > superset.len() {
+        return false;
+    }
+
+    let mut superset = superset.iter();
+    for expected in subset {
+        loop {
+            match superset.next() {
+                Some(candidate) if candidate < expected => {}
+                Some(candidate) if candidate == expected => break,
+                _ => return false,
+            }
+        }
+    }
+    true
 }
 
 /// Normalizes one boolean expression into an equivalent, solver-friendlier form.
@@ -40,6 +85,56 @@ pub(crate) fn normalize_bool_for_solver(expr: SymBoolExpr) -> SymBoolExpr {
 }
 
 impl SymBoolExpr {
+    fn cache_key(self) -> Self {
+        self.fold(&mut Self::cache_key_node)
+    }
+
+    fn cache_key_node(expr: Self) -> Self {
+        match expr.into_kind() {
+            SymBoolExprKind::Not(value) => value.not(),
+            SymBoolExprKind::And(values) => {
+                let mut conjuncts = Vec::new();
+                for value in values.iter().cloned() {
+                    value.push_cache_key_conjuncts(&mut conjuncts);
+                }
+                sort_dedup_bool_exprs(&mut conjuncts);
+                Self::and(conjuncts)
+            }
+            SymBoolExprKind::Eq(left, right) => {
+                let left = left.cache_key();
+                let right = right.cache_key();
+                if left <= right { Self::eq(left, right) } else { Self::eq(right, left) }
+            }
+            SymBoolExprKind::Cmp(op, left, right) => {
+                Self::cache_key_cmp(op, left.cache_key(), right.cache_key())
+            }
+            SymBoolExprKind::Const(value) => Self::constant(value),
+        }
+    }
+
+    fn push_cache_key_conjuncts(self, out: &mut Vec<Self>) {
+        match self.kind() {
+            SymBoolExprKind::Const(true) => {}
+            SymBoolExprKind::And(values) => {
+                for value in values.iter().cloned() {
+                    value.push_cache_key_conjuncts(out);
+                }
+            }
+            _ => out.push(self),
+        }
+    }
+
+    fn cache_key_cmp(op: SymBoolExprOp, left: SymExpr, right: SymExpr) -> Self {
+        match op {
+            SymBoolExprOp::Ugt => Self::cmp(SymBoolExprOp::Ult, right, left),
+            SymBoolExprOp::Uge => Self::cmp(SymBoolExprOp::Ule, right, left),
+            SymBoolExprOp::Sgt => Self::cmp(SymBoolExprOp::Slt, right, left),
+            SymBoolExprOp::Ult | SymBoolExprOp::Ule | SymBoolExprOp::Slt => {
+                Self::cmp(op, left, right)
+            }
+        }
+    }
+
     fn push_normalized_conjuncts(self, out: &mut Vec<Self>) {
         match self.kind() {
             SymBoolExprKind::Const(true) => {}
@@ -50,6 +145,421 @@ impl SymBoolExpr {
             }
             _ => out.push(self),
         }
+    }
+}
+
+impl SymExpr {
+    fn cache_key(self) -> Self {
+        self.fold(&mut Self::cache_key_node)
+    }
+
+    fn cache_key_node(expr: Self) -> Self {
+        match expr.kind() {
+            SymExprKind::Op(op, left, right) => {
+                if op.is_commutative() && right < left {
+                    let SymExprKind::Op(op, left, right) = expr.into_kind() else { unreachable!() };
+                    Self::op(op, right, left)
+                } else {
+                    expr
+                }
+            }
+            SymExprKind::AddMod { left, right, .. } => {
+                if right < left {
+                    let SymExprKind::AddMod { left, right, modulus } = expr.into_kind() else {
+                        unreachable!()
+                    };
+                    Self::addmod(right, left, modulus)
+                } else {
+                    expr
+                }
+            }
+            SymExprKind::MulMod { left, right, .. } => {
+                if right < left {
+                    let SymExprKind::MulMod { left, right, modulus } = expr.into_kind() else {
+                        unreachable!()
+                    };
+                    Self::mulmod(right, left, modulus)
+                } else {
+                    expr
+                }
+            }
+            SymExprKind::Ite(_, _, _) => {
+                let SymExprKind::Ite(cond, left, right) = expr.into_kind() else { unreachable!() };
+                Self::ite(cond.cache_key(), left, right)
+            }
+            _ => expr,
+        }
+    }
+}
+
+impl SymExprOp {
+    const fn is_commutative(self) -> bool {
+        matches!(self, Self::Add | Self::Mul | Self::And | Self::Or | Self::Xor)
+    }
+}
+
+pub(super) fn write_smt_assertions(out: &mut String, constraints: &[SymBoolExpr]) {
+    if constraints.is_empty() {
+        return;
+    }
+
+    let plan = SmtCsePlan::new(constraints);
+    if plan.bindings.is_empty() {
+        for constraint in constraints {
+            let _ = writeln!(out, "(assert {})", constraint.smt());
+        }
+        return;
+    }
+
+    let writer = SmtCseWriter { plan: &plan };
+    // assert:
+    //   let binding_0 = term_0
+    //   ...
+    //   let binding_n = term_n
+    //   and(constraints)
+    out.push_str("(assert ");
+    for (idx, binding) in plan.bindings.iter().enumerate() {
+        out.push_str("(let ((");
+        binding.write_name(out, idx);
+        out.push(' ');
+        match binding {
+            SmtBinding::Expr(expr) => writer.write_expr(out, expr, Some(idx), None),
+            SmtBinding::Bool(expr) => writer.write_bool(out, expr, None, Some(idx)),
+        }
+        out.push_str(")) ");
+    }
+    writer.write_bool_conjunction(out, constraints);
+    for _ in &plan.bindings {
+        out.push(')');
+    }
+    out.push_str(")\n");
+}
+
+#[derive(Default)]
+struct SmtCseVisit {
+    count: usize,
+    binding: Option<usize>,
+}
+
+struct SmtCsePlan {
+    expr_visits: HashMap<SymExpr, SmtCseVisit>,
+    bool_visits: HashMap<SymBoolExpr, SmtCseVisit>,
+    bindings: Vec<SmtBinding>,
+}
+
+impl SmtCsePlan {
+    fn new(constraints: &[SymBoolExpr]) -> Self {
+        let mut plan = Self {
+            expr_visits: HashMap::default(),
+            bool_visits: HashMap::default(),
+            bindings: Vec::new(),
+        };
+        for constraint in constraints {
+            plan.count_bool(constraint);
+        }
+        for constraint in constraints {
+            plan.collect_bool_binding(constraint);
+        }
+        plan
+    }
+
+    fn count_expr(&mut self, expr: &SymExpr) {
+        self.expr_visits.entry(expr.clone()).or_default().count += 1;
+        match expr.kind() {
+            SymExprKind::Const(_)
+            | SymExprKind::Var(_)
+            | SymExprKind::GasLeft(_)
+            | SymExprKind::Keccak { .. }
+            | SymExprKind::Hash { .. } => {}
+            SymExprKind::Not(value) => self.count_expr(value),
+            SymExprKind::Op(_, left, right) => {
+                self.count_expr(left);
+                self.count_expr(right);
+            }
+            SymExprKind::AddMod { left, right, modulus }
+            | SymExprKind::MulMod { left, right, modulus } => {
+                self.count_expr(modulus);
+                self.count_expr(left);
+                self.count_expr(right);
+                self.count_expr(modulus);
+            }
+            SymExprKind::Ite(cond, left, right) => {
+                self.count_bool(cond);
+                self.count_expr(left);
+                self.count_expr(right);
+            }
+        }
+    }
+
+    fn count_bool(&mut self, expr: &SymBoolExpr) {
+        self.bool_visits.entry(expr.clone()).or_default().count += 1;
+        match expr.kind() {
+            SymBoolExprKind::Const(_) => {}
+            SymBoolExprKind::Not(value) => self.count_bool(value),
+            SymBoolExprKind::And(values) => {
+                for value in values.iter() {
+                    self.count_bool(value);
+                }
+            }
+            SymBoolExprKind::Eq(left, right) | SymBoolExprKind::Cmp(_, left, right) => {
+                self.count_expr(left);
+                self.count_expr(right);
+            }
+        }
+    }
+
+    fn collect_expr_binding(&mut self, expr: &SymExpr) {
+        match expr.kind() {
+            SymExprKind::Const(_)
+            | SymExprKind::Var(_)
+            | SymExprKind::GasLeft(_)
+            | SymExprKind::Keccak { .. }
+            | SymExprKind::Hash { .. } => {}
+            SymExprKind::Not(value) => self.collect_expr_binding(value),
+            SymExprKind::Op(_, left, right) => {
+                self.collect_expr_binding(left);
+                self.collect_expr_binding(right);
+            }
+            SymExprKind::AddMod { left, right, modulus }
+            | SymExprKind::MulMod { left, right, modulus } => {
+                self.collect_expr_binding(modulus);
+                self.collect_expr_binding(left);
+                self.collect_expr_binding(right);
+            }
+            SymExprKind::Ite(cond, left, right) => {
+                self.collect_bool_binding(cond);
+                self.collect_expr_binding(left);
+                self.collect_expr_binding(right);
+            }
+        }
+        self.bind_expr(expr);
+    }
+
+    fn collect_bool_binding(&mut self, expr: &SymBoolExpr) {
+        match expr.kind() {
+            SymBoolExprKind::Const(_) => {}
+            SymBoolExprKind::Not(value) => self.collect_bool_binding(value),
+            SymBoolExprKind::And(values) => {
+                for value in values.iter() {
+                    self.collect_bool_binding(value);
+                }
+            }
+            SymBoolExprKind::Eq(left, right) | SymBoolExprKind::Cmp(_, left, right) => {
+                self.collect_expr_binding(left);
+                self.collect_expr_binding(right);
+            }
+        }
+        self.bind_bool(expr);
+    }
+
+    fn bind_expr(&mut self, expr: &SymExpr) {
+        let Some(visit) = self.expr_visits.get_mut(expr) else { return };
+        if visit.count <= 1 || visit.binding.is_some() || !Self::expr_can_bind(expr) {
+            return;
+        }
+        let idx = self.bindings.len();
+        visit.binding = Some(idx);
+        self.bindings.push(SmtBinding::Expr(expr.clone()));
+    }
+
+    fn bind_bool(&mut self, expr: &SymBoolExpr) {
+        let Some(visit) = self.bool_visits.get_mut(expr) else { return };
+        if visit.count <= 1 || visit.binding.is_some() || !Self::bool_can_bind(expr) {
+            return;
+        }
+        let idx = self.bindings.len();
+        visit.binding = Some(idx);
+        self.bindings.push(SmtBinding::Bool(expr.clone()));
+    }
+
+    fn expr_binding(&self, expr: &SymExpr) -> Option<usize> {
+        self.expr_visits.get(expr).and_then(|visit| visit.binding)
+    }
+
+    fn bool_binding(&self, expr: &SymBoolExpr) -> Option<usize> {
+        self.bool_visits.get(expr).and_then(|visit| visit.binding)
+    }
+
+    fn expr_can_bind(expr: &SymExpr) -> bool {
+        !matches!(
+            expr.kind(),
+            SymExprKind::Const(_)
+                | SymExprKind::Var(_)
+                | SymExprKind::GasLeft(_)
+                | SymExprKind::Keccak { .. }
+                | SymExprKind::Hash { .. }
+        )
+    }
+
+    fn bool_can_bind(expr: &SymBoolExpr) -> bool {
+        !matches!(expr.kind(), SymBoolExprKind::Const(_))
+    }
+}
+
+enum SmtBinding {
+    Expr(SymExpr),
+    Bool(SymBoolExpr),
+}
+
+impl SmtBinding {
+    fn write_name(&self, out: &mut String, idx: usize) {
+        match self {
+            Self::Expr(_) => Self::write_expr_name(out, idx),
+            Self::Bool(_) => Self::write_bool_name(out, idx),
+        }
+    }
+
+    fn write_expr_name(out: &mut String, idx: usize) {
+        let _ = write!(out, "__sym_expr_{idx}");
+    }
+
+    fn write_bool_name(out: &mut String, idx: usize) {
+        let _ = write!(out, "__sym_bool_{idx}");
+    }
+}
+
+struct SmtCseWriter<'a> {
+    plan: &'a SmtCsePlan,
+}
+
+impl SmtCseWriter<'_> {
+    fn write_expr(
+        &self,
+        out: &mut String,
+        expr: &SymExpr,
+        skip_expr: Option<usize>,
+        skip_bool: Option<usize>,
+    ) {
+        if let Some(idx) = self.plan.expr_binding(expr)
+            && Some(idx) != skip_expr
+        {
+            SmtBinding::write_expr_name(out, idx);
+            return;
+        }
+
+        match expr.kind() {
+            SymExprKind::Const(value) => {
+                let _ = write!(out, "(_ bv{value} 256)");
+            }
+            SymExprKind::Var(var) => out.push_str(var.as_str()),
+            SymExprKind::GasLeft(id) => {
+                let _ = write!(out, "gasleft_{id}");
+            }
+            SymExprKind::Keccak { name, .. } => out.push_str(name.as_str()),
+            SymExprKind::Hash { name, .. } => out.push_str(name.as_str()),
+            SymExprKind::Not(value) => {
+                out.push_str("(bvnot ");
+                self.write_expr(out, value, skip_expr, skip_bool);
+                out.push(')');
+            }
+            SymExprKind::Op(op, left, right) => {
+                let _ = write!(out, "({} ", op.smt());
+                self.write_expr(out, left, skip_expr, skip_bool);
+                out.push(' ');
+                self.write_expr(out, right, skip_expr, skip_bool);
+                out.push(')');
+            }
+            SymExprKind::AddMod { left, right, modulus } => {
+                self.write_wide_modular_arithmetic(out, "bvadd", left, right, modulus);
+            }
+            SymExprKind::MulMod { left, right, modulus } => {
+                self.write_wide_modular_arithmetic(out, "bvmul", left, right, modulus);
+            }
+            SymExprKind::Ite(cond, left, right) => {
+                out.push_str("(ite ");
+                self.write_bool(out, cond, skip_expr, skip_bool);
+                out.push(' ');
+                self.write_expr(out, left, skip_expr, skip_bool);
+                out.push(' ');
+                self.write_expr(out, right, skip_expr, skip_bool);
+                out.push(')');
+            }
+        }
+    }
+
+    fn write_wide_modular_arithmetic(
+        &self,
+        out: &mut String,
+        op: &'static str,
+        left: &SymExpr,
+        right: &SymExpr,
+        modulus: &SymExpr,
+    ) {
+        // if modulus == 0:
+        //   0
+        // else:
+        //   low_256((zext(left) op zext(right)) urem zext(modulus))
+        out.push_str("(ite (= ");
+        self.write_expr(out, modulus, None, None);
+        out.push_str(" (_ bv0 256)) (_ bv0 256) ((_ extract 255 0) (bvurem (");
+        out.push_str(op);
+        out.push_str(" ((_ zero_extend 256) ");
+        self.write_expr(out, left, None, None);
+        out.push_str(") ((_ zero_extend 256) ");
+        self.write_expr(out, right, None, None);
+        out.push_str(")) ((_ zero_extend 256) ");
+        self.write_expr(out, modulus, None, None);
+        out.push_str("))))");
+    }
+
+    fn write_bool(
+        &self,
+        out: &mut String,
+        expr: &SymBoolExpr,
+        skip_expr: Option<usize>,
+        skip_bool: Option<usize>,
+    ) {
+        if let Some(idx) = self.plan.bool_binding(expr)
+            && Some(idx) != skip_bool
+        {
+            SmtBinding::write_bool_name(out, idx);
+            return;
+        }
+
+        match expr.kind() {
+            SymBoolExprKind::Const(value) => out.push_str(if *value { "true" } else { "false" }),
+            SymBoolExprKind::Not(value) => {
+                out.push_str("(not ");
+                self.write_bool(out, value, skip_expr, skip_bool);
+                out.push(')');
+            }
+            SymBoolExprKind::And(values) => {
+                out.push_str("(and");
+                for value in values.iter() {
+                    out.push(' ');
+                    self.write_bool(out, value, skip_expr, skip_bool);
+                }
+                out.push(')');
+            }
+            SymBoolExprKind::Eq(left, right) => {
+                out.push_str("(= ");
+                self.write_expr(out, left, skip_expr, skip_bool);
+                out.push(' ');
+                self.write_expr(out, right, skip_expr, skip_bool);
+                out.push(')');
+            }
+            SymBoolExprKind::Cmp(op, left, right) => {
+                let _ = write!(out, "({} ", op.smt());
+                self.write_expr(out, left, skip_expr, skip_bool);
+                out.push(' ');
+                self.write_expr(out, right, skip_expr, skip_bool);
+                out.push(')');
+            }
+        }
+    }
+
+    fn write_bool_conjunction(&self, out: &mut String, constraints: &[SymBoolExpr]) {
+        if constraints.len() == 1 {
+            self.write_bool(out, &constraints[0], None, None);
+            return;
+        }
+
+        out.push_str("(and");
+        for constraint in constraints {
+            out.push(' ');
+            self.write_bool(out, constraint, None, None);
+        }
+        out.push(')');
     }
 }
